backendv2/requirements.txt
Django==4.2
DjangoRestFramework
Django-Cors-Headers
Requests